Add an "rm -rf" to the suggested fix-up commmand.
The git-svn-clone-externals script has the shortcoming that it doesn't
do the right thing when invoked on a directory that it was already used
to setup in the past.  Blowing it away and running it again fixes
things, although it might not always be what you want if the externals
directory is also being used for development.