Use Digest::MD5 instead of crypt for cache file names.
This bug is actually pretty shameful.  I was just looking for a
convenient hash function and used crypt without thinking.  Unfortunately
crypt only hashes the first 8 characters of the string, so if you store
all your git repositories in the same directory then you could be sad
when you get a hash conflict.  MD5 ought to work much better for this
purpose.